// src/app/core/interfaces/cita.module.ts
export interface Cita {
  cita_id: number;
  cita_usuario_id: number;
  cita_usuario_nombre: string;
  cita_profesional_id: number;
  cita_profesional_nombre: string;
  cita_tipo_id: number;
  cita_tipo_nombre: string;
  cita_fecha: string;
  cita_estado_id: number;
  cita_estado_nombre: string;
  cita_resultado: string;
}